redesign of the SortStack and SortStore classes: created a WeakPriorityBlockingQueue as special implementation of a PriorityBlockingQueue with a weak object binding. - better abstraction of ordering technique - fixed some bugs according to result numbering (distinguish different counters in Queue) - fixed a ordering bug in post-ranking (ordering was decreased instead of increased) - reversed ordering numbering using a reversed ordering. The higher the ranking number the better (now).
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@7128 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/search/MediaSnippet.java
@@ -49,11 +49,11 @@ public class MediaSnippet implements Comparable<MediaSnippet>, Comparator<MediaS
     public ContentDomain type;
     public DigestURI href, source;
     public String name, attr, mime;
-    public int ranking;
+    public long ranking;
     public int width, height;
     public long fileSize;
 
-    public MediaSnippet(final ContentDomain type, final DigestURI href, final String mime, final String name, final long fileSize, final String attr, final int ranking, final DigestURI source) {
+    public MediaSnippet(final ContentDomain type, final DigestURI href, final String mime, final String name, final long fileSize, final String attr, final long ranking, final DigestURI source) {
         this.type = type;
         this.href = href;
         this.mime = mime;
@@ -73,7 +73,7 @@ public class MediaSnippet implements Comparable<MediaSnippet>, Comparator<MediaS
         if ((this.attr == null) || (this.attr.length() == 0)) this.attr = "_";
     }
     
-    public MediaSnippet(final ContentDomain type, final DigestURI href, final String mime, final String name, final long fileSize, final int width, final int height, final int ranking, final DigestURI source) {
+    public MediaSnippet(final ContentDomain type, final DigestURI href, final String mime, final String name, final long fileSize, final int width, final int height, final long ranking, final DigestURI source) {
         this.type = type;
         this.href = href;
         this.mime = mime;
@@ -188,7 +188,7 @@ public class MediaSnippet implements Comparable<MediaSnippet>, Comparator<MediaS
             int appcount = queryhashes.size()  * 2 - 
                            TextSnippet.removeAppearanceHashes(url.toNormalform(false, false), queryhashes).size() -
                            TextSnippet.removeAppearanceHashes(desc, queryhashes).size();
-            final int ranking = Integer.MAX_VALUE - (ientry.height() + 1) * (ientry.width() + 1) * (appcount + 1);  
+            final long ranking = Long.MAX_VALUE - (ientry.height() + 1) * (ientry.width() + 1) * (appcount + 1);  
             result.add(new MediaSnippet(ContentDomain.IMAGE, url, MimeTable.url2mime(url), desc, ientry.fileSize(), ientry.width(), ientry.height(), ranking, source));
         }
         return result;
